requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.txt
#
alembic==1.8.1
    # via
    #   -r requirements.txt
    #   flask-migrate
babel==2.11.0
    # via
    #   -r requirements.txt
    #   flask-babel
blinker==1.5
    # via
    #   -r requirements.txt
    #   flask-mail
build==0.9.0
    # via
    #   -r requirements.txt
    #   pip-tools
certifi==2022.12.7
    # via
    #   -r requirements.txt
    #   requests
charset-normalizer==2.1.1
    # via
    #   -r requirements.txt
    #   requests
click==8.1.3
    # via
    #   -r requirements.txt
    #   flask
    #   pip-tools
dnspython==2.2.1
    # via
    #   -r requirements.txt
    #   email-validator
dominate==2.7.0
    # via
    #   -r requirements.txt
    #   flask-bootstrap
email-validator==1.3.0
    # via -r requirements.txt
flake8==6.0.0
    # via -r requirements.txt
flask==2.2.2
    # via
    #   -r requirements.txt
    #   flask-babel
    #   flask-bootstrap
    #   flask-login
    #   flask-mail
    #   flask-migrate
    #   flask-moment
    #   flask-sqlalchemy
    #   flask-wtf
flask-babel==2.0.0
    # via -r requirements.txt
flask-bootstrap==3.3.7.1
    # via -r requirements.txt
flask-login==0.6.2
    # via -r requirements.txt
flask-mail==0.9.1
    # via -r requirements.txt
flask-migrate==4.0.0
    # via -r requirements.txt
flask-moment==1.0.5
    # via -r requirements.txt
flask-sqlalchemy==3.0.2
    # via
    #   -r requirements.txt
    #   flask-migrate
flask-wtf==1.0.1
    # via -r requirements.txt
greenlet==2.0.1
    # via
    #   -r requirements.txt
    #   sqlalchemy
guess-language-spirit==0.5.3
    # via -r requirements.txt
idna==3.4
    # via
    #   -r requirements.txt
    #   email-validator
    #   requests
itsdangerous==2.1.2
    # via
    #   -r requirements.txt
    #   flask
    #   flask-wtf
jinja2==3.1.2
    # via
    #   -r requirements.txt
    #   flask
    #   flask-babel
mako==1.2.4
    # via
    #   -r requirements.txt
    #   alembic
markupsafe==2.1.1
    # via
    #   -r requirements.txt
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
mccabe==0.7.0
    # via
    #   -r requirements.txt
    #   flake8
packaging==22.0
    # via
    #   -r requirements.txt
    #   build
    #   flask-moment
pep517==0.13.0
    # via
    #   -r requirements.txt
    #   build
pip-tools==6.12.0
    # via -r requirements.txt
psycopg2==2.9.5
    # via -r requirements.txt
pycodestyle==2.10.0
    # via
    #   -r requirements.txt
    #   flake8
pyflakes==3.0.1
    # via
    #   -r requirements.txt
    #   flake8
pyjwt==2.6.0
    # via -r requirements.txt
python-dotenv==0.21.0
    # via -r requirements.txt
pytz==2022.6
    # via
    #   -r requirements.txt
    #   babel
    #   flask-babel
requests==2.28.1
    # via -r requirements.txt
sqlalchemy==1.4.44
    # via
    #   -r requirements.txt
    #   alembic
    #   flask-sqlalchemy
tomli==2.0.1
    # via
    #   -r requirements.txt
    #   build
    #   pep517
urllib3==1.26.13
    # via
    #   -r requirements.txt
    #   requests
visitor==0.1.3
    # via
    #   -r requirements.txt
    #   flask-bootstrap
werkzeug==2.2.2
    # via
    #   -r requirements.txt
    #   flask
    #   flask-login
wheel==0.38.4
    # via
    #   -r requirements.txt
    #   pip-tools
wtforms==3.0.1
    # via
    #   -r requirements.txt
    #   flask-wtf

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
